test(azure): split batch api tests
